Update navX libraries
* Fix I2C write method
* Update for WPILib LiveWindow changes
* Fix decoding of GyroFSRDPS and AccelFSRG

This intentionally does not implement the enableLogging methods,
as these can be filtered out by configuring the logging module.

It appears the I2C read method was fixed for the flipped WPILib I2C
write return values in 2017, but the write method wasn't fixed, as
RegisterIO expects that a return value of True from write is success.